+++ src/main/java/algorithms/BPNN/BPNN.java
@@ -21,8 +21,8 @@ public class BPNN {
     private static double w45 = 1.1;
     private static double rateOfEvalution = 0.7;
     private static int count = 10000;
-    final static int inputs[][] = { {1, 1}, {1, 0}, {0, 1}, {0, 0}};
-    final static int expectedOutputs[] = {0, 1, 1, 0};
+    final static double inputs[][] = { {1, 1}, {1, 0}, {0, 1}, {0, 0}};
+    final static double expectedOutputs[] = {0, 1, 1, 0};
 
     public static void main(String[] args) {
         NeuralNetwork network = init(rateOfEvalution);
